+++ src/main/java/org/goafabric/personservice/controller/vo/Person.java
@@ -1,15 +1,16 @@
 package org.goafabric.personservice.controller.vo;
 
-import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.Null;
-import jakarta.validation.constraints.Size;
-
 import java.util.List;
+import java.util.Objects;
 
 public record Person (
-    @Null String id,
-    @Null String version,
-    @NotNull @Size(min = 3, max = 255) String firstName,
-    @NotNull @Size(min = 3, max = 255) String lastName,
-    List<Address> address
-) {}
+    String id,
+    String version,
+    String firstName,
+    String lastName,
+    List<Address> address) {
+    public Person {
+        Objects.requireNonNull(firstName);
+        Objects.requireNonNull(lastName);
+    }
+}