+++ doc/release-notes.md
@@ -5,6 +5,7 @@
 
 - one to many relation added
 - declarative webclient replaced again by resttemplate fo future declarative restclient
+- validation framework removed
 
 # 3.1.1
 - upgrade to Spring Boot 3.1.1